@@ -4,37 +4,37 @@
   "private": true,
   "proxy": "http://localhost:3001",
   "dependencies": {
-    "@apollo/client": "^3.4.17",
+    "@apollo/client": "^3.7.3",
     "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/apollo-upload-client": "^14.1.0",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.7",
-    "@types/react": "^17.0.34",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-router-dom": "^5.3.2",
-    "apollo-upload-client": "^16.0.0",
-    "cypress": "^9.0.0",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^19.0.0",
-    "eslint-config-airbnb-typescript": "^15.0.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.3",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/apollo-upload-client": "^17.0.2",
+    "@types/jest": "^29.2.5",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "@types/react-router-dom": "^5.3.3",
+    "apollo-upload-client": "^17.0.0",
+    "cypress": "^12.2.0",
+    "eslint": "^8.30.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.27.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "graphql": "^16.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-infinite-scroll-hook": "^4.0.1",
-    "react-query": "^3.22.0",
-    "react-responsive-carousel": "^3.2.21",
-    "react-router-dom": "^5.2.1",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.4.4",
-    "web-vitals": "^1.1.2"
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "graphql": "^16.6.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-infinite-scroll-hook": "^4.0.4",
+    "react-query": "^3.39.2",
+    "react-responsive-carousel": "^3.2.23",
+    "react-router-dom": "^6.6.1",
+    "react-scripts": "^5.0.1",
+    "typescript": "^4.9.4",
+    "web-vitals": "^3.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -58,8 +58,8 @@
     ]
   },
   "devDependencies": {
-    "lint-staged": "^11.2.6",
-    "prettier": "^2.4.1"
+    "lint-staged": "^13.1.0",
+    "prettier": "^2.8.1"
   },
   "lint-staged": {
     "src/**/*.tsx": [
