@@ -8,55 +8,54 @@
     "dev": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/index.ts",
     "build": "tsc",
     "production": "node dist/src/index.js",
-    "test": "jest",
-    "precommit": "lint-staged",
-    "prepush": "jest --passWithNoTests"
+    "test": "jest"
   },
   "author": "williamstyronejr",
   "license": "ISC",
   "dependencies": {
+    "@apollo/server": "^4.3.0",
     "@types/bcrypt": "^5.0.0",
-    "@types/cookie-parser": "^1.4.2",
+    "@types/cookie-parser": "^1.4.3",
+    "@types/cors": "^2.8.13",
     "@types/csurf": "^1.11.2",
-    "@types/express": "^4.17.13",
-    "@types/graphql-upload": "^8.0.7",
-    "@types/jest": "^27.0.2",
-    "@types/jsonwebtoken": "^8.5.5",
-    "@types/node": "^16.11.7",
-    "@types/redis": "^2.8.32",
+    "@types/express": "^4.17.15",
+    "@types/jest": "^29.2.4",
+    "@types/jsonwebtoken": "^9.0.0",
+    "@types/node": "^18.11.18",
+    "@types/redis": "^4.0.11",
     "@types/validator": "^13.6.6",
-    "@typescript-eslint/eslint-plugin": "^5.3.1",
-    "@typescript-eslint/parser": "^5.3.1",
-    "apollo-server": "^3.5.0",
-    "apollo-server-core": "^3.5.0",
-    "apollo-server-express": "^3.5.0",
-    "bcrypt": "^5.0.1",
-    "cookie-parser": "^1.4.5",
+    "@typescript-eslint/eslint-plugin": "^5.47.1",
+    "@typescript-eslint/parser": "^5.47.1",
+    "bcrypt": "^5.1.0",
+    "body-parser": "^1.20.1",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "csurf": "^1.11.0",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "graphql": "^15.7.2",
-    "graphql-upload": "^12.0.0",
-    "jsonwebtoken": "^8.5.1",
-    "passport": "^0.4.1",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
+    "graphql": "^16.6.0",
+    "graphql-tag": "^2.12.6",
+    "graphql-upload": "^15.0.1",
+    "jsonwebtoken": "^9.0.0",
+    "passport": "^0.6.0",
     "passport-google-oauth": "^2.0.0",
-    "pg": "^8.7.1",
+    "pg": "^8.8.0",
     "pg-hstore": "^2.3.4",
-    "redis": "^3.1.2",
-    "sequelize": "^6.9.0",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.4.4",
-    "winston": "^3.3.3"
+    "redis": "^4.5.1",
+    "sequelize": "^6.28.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4",
+    "winston": "^3.8.2"
   },
   "devDependencies": {
+    "@types/graphql-upload": "^15.0.0",
     "babel": "^6.23.0",
-    "eslint": "^8.2.0",
-    "jest": "^27.3.1",
-    "lint-staged": "^11.2.6",
-    "nodemon": "^2.0.15",
-    "prettier": "^2.4.1",
-    "ts-jest": "^27.0.7"
+    "eslint": "^8.30.0",
+    "jest": "^29.3.1",
+    "lint-staged": "^13.1.0",
+    "nodemon": "^2.0.20",
+    "prettier": "^2.8.1",
+    "ts-jest": "^29.0.3"
   },
   "lint-staged": {
     "src/**/*.ts": [
