@@ -2,8 +2,8 @@ import * as React from 'react';
 import {
   BrowserRouter as Router,
   Route,
-  Switch,
-  Redirect,
+  Routes,
+  Navigate,
 } from 'react-router-dom';
 import HomePage from './scenes/Home';
 import Header from './components/Header';
@@ -30,10 +30,10 @@ import RecoveryPage from './scenes/auth/Recovery';
 import FavoritesPage from './scenes/dashboard/Favorites';
 import { AuthProvider, useAuthContext } from './context/auth';
 
-const ProtectedRoutes = ({ children }: { children: any }) => {
+// eslint-disable-next-line react/require-default-props
+const ProtectedRoutes: React.FC<{ children?: any }> = ({ children = null }) => {
   const { state } = useAuthContext();
-
-  return state.authenticated ? children : <Redirect to="/signin" />;
+  return state.authenticated ? children : <Navigate to="/signin" />;
 };
 
 export default () => (
@@ -41,37 +41,87 @@ export default () => (
     <AuthProvider>
       <Header />
       <main className="main">
-        <Switch>
-          <Route exact path="/" component={HomePage} />
-          <Route path="/faq" component={FaqPage} />
-          <Route path="/started" component={StartedPage} />
-          <Route path="/feedback" component={FeedbackPage} />
-          <Route path="/about" component={AboutPage} />
-          <Route path="/support" component={SupportPage} />
+        <Routes>
+          <Route path="/" element={<HomePage />} />
+          <Route path="/faq" element={<FaqPage />} />
+          <Route path="/started" element={<StartedPage />} />
+          <Route path="/feedback" element={<FeedbackPage />} />
+          <Route path="/about" element={<AboutPage />} />
+          <Route path="/support" element={<SupportPage />} />
 
-          <Route path="/signin" component={SigninPage} />
-          <Route path="/signup" component={SignupPage} />
-          <Route path="/recovery" component={RecoveryPage} />
+          <Route path="/signin" element={<SigninPage />} />
+          <Route path="/signup" element={<SignupPage />} />
+          <Route path="/recovery" element={<RecoveryPage />} />
 
-          <Route path="/search" component={SearchPage} />
-          <Route exact path="/recipe/:recipeId" component={RecipePage} />
-          <Route path="/explore/:type" component={ExplorePage} />
+          <Route path="/search" element={<SearchPage />} />
+          <Route path="/recipe/:recipeId" element={<RecipePage />} />
+          <Route path="/explore/:type" element={<ExplorePage />} />
 
-          <Route path="/account/profile/:userId" component={ProfilePage} />
-          <ProtectedRoutes>
-            <Route path="/recipe/:recipeId/edit" component={EditRecipePage} />
+          <Route path="/account/profile/:userId" element={<ProfilePage />} />
 
-            <Route exact path="/dashboard" component={DashboardPage} />
-            <Route path="/dashboard/manage" component={ManageRecipesPage} />
-            <Route path="/dashboard/analytics" component={AnalyticsPage} />
-            <Route path="/dashboard/favorites" component={FavoritesPage} />
+          <Route
+            path="/recipe/:recipeId/edit"
+            element={
+              <ProtectedRoutes>
+                <EditRecipePage />
+              </ProtectedRoutes>
+            }
+          />
 
-            <Route path="/account/settings" component={SettingsPage} />
-            <Route path="/account/recipe/create" component={CreateRecipePage} />
-          </ProtectedRoutes>
-          <Route component={MissingPage} />
-        </Switch>
+          <Route
+            path="/dashboard"
+            element={
+              <ProtectedRoutes>
+                <DashboardPage />
+              </ProtectedRoutes>
+            }
+          />
+          <Route
+            path="/dashboard/manage"
+            element={
+              <ProtectedRoutes>
+                <ManageRecipesPage />
+              </ProtectedRoutes>
+            }
+          />
+          <Route
+            path="/dashboard/analytics"
+            element={
+              <ProtectedRoutes>
+                <AnalyticsPage />
+              </ProtectedRoutes>
+            }
+          />
+          <Route
+            path="/dashboard/favorites"
+            element={
+              <ProtectedRoutes>
+                <FavoritesPage />
+              </ProtectedRoutes>
+            }
+          />
+
+          <Route
+            path="/account/settings"
+            element={
+              <ProtectedRoutes>
+                <SettingsPage />
+              </ProtectedRoutes>
+            }
+          />
+          <Route
+            path="/account/recipe/create"
+            element={
+              <ProtectedRoutes>
+                <CreateRecipePage />
+              </ProtectedRoutes>
+            }
+          />
+
+          <Route element={<MissingPage />} />
+        </Routes>
       </main>
+
       <Footer />
     </AuthProvider>
   </Router>
