@@ -1,6 +1,13 @@
-import { AuthenticationError } from 'apollo-server';
 import jwt from 'jsonwebtoken';
+import { GraphQLError } from 'graphql';
 import logger from '../../services/logger';
+
+const AuthenticationError = (msg: string) =>
+  new GraphQLError(msg, {
+    extensions: {
+      code: 'UNAUTHENTICATED',
+    },
+  });
 
 const { JWT_SECRET } = process.env;
 
@@ -14,9 +21,9 @@ export default function checkAuth(context: any, required = true): any {
     } catch (err) {
       logger.info('User cookie was invalid');
       context.res.clearCookie('token');
-      if (required) throw new AuthenticationError('Invalid/Expired token');
+      if (required) throw AuthenticationError('Invalid/Expired token');
     }
   }
 
-  if (required) throw new AuthenticationError('No token was found.');
+  if (required) throw AuthenticationError('No token was found.');
 }
